Move api/server/router/image to daemon/server/router/image
