+++ daemon/command/daemon.go
@@ -23,7 +23,6 @@ import (
 	buildbackend "github.com/docker/docker/api/server/backend/build"
 	"github.com/docker/docker/api/server/middleware"
 	"github.com/docker/docker/api/server/router"
-	"github.com/docker/docker/api/server/router/image"
 	"github.com/docker/docker/api/server/router/network"
 	pluginrouter "github.com/docker/docker/api/server/router/plugin"
 	sessionrouter "github.com/docker/docker/api/server/router/session"
@@ -45,6 +44,7 @@ import (
 	debugrouter "github.com/docker/docker/daemon/server/router/debug"
 	distributionrouter "github.com/docker/docker/daemon/server/router/distribution"
 	grpcrouter "github.com/docker/docker/daemon/server/router/grpc"
+	"github.com/docker/docker/daemon/server/router/image"
 	"github.com/docker/docker/dockerversion"
 	"github.com/docker/docker/internal/otelutil"
 	"github.com/docker/docker/libcontainerd/supervisor"
