+++ daemon/server/router/image/image_routes.go
