+++ daemon/server/router/image/inspect_response.go
