+++ daemon/server/router/image/inspect_response_test.go
